TraceDR-model/requirements.txt
torch>=1.8.0
transformers>=4.0.0
py2neo>=2021.0
jieba>=0.42.0
scikit-learn>=0.24.0
python-Levenshtein>=0.12.0
tqdm>=4.60.0
numpy>=1.20.0
pandas>=1.3.0
neo4j>=4.4.0 